Reset milestone selection upon repo change (#57)
Previously, milestones selected in the milestone filter are not cleared 
upon a repository change. This means that old milestones from the 
previous repository remain selected, which may or may not exist in the 
newly loaded repository. As such, the milestone selection should be 
fresh and reset for every new repo loaded.

Let's reset the milestone selection upon a repository change