Fix pressing back to cause issues to being cleared (#142)
Issues stored in issue service are being reset each time initialize is
called in IssuesViewerComponent. This is undesirable as this method
can be called when users manually press back on the browser. 

As the issues cache is being managed via a cache manager, this will
create a scenario where the cache manager claims that there are no
changes to the repo but the cache has already been cleared. This causes
issues to be cleared and no new issues are being loaded from the GitHub
Server.

Let's resolve this by moving the reset to be called when repo change
event occurs instead.