Disable milestone filter if there are no milestones (#149)
Previously, when there are no milestones in a repo, there is still a 
`Milestone.DefaultMilestone` instance in WATcher, with the string 
representation of `Without a milestone` for issues without a milestone. 
This causes the milestone selector to show `Without a milestone` as the 
trigger text when it is selected.

Let's
* Disable the milestone selector
* Set the display text to `No Milestones`
when there are no milestones in a repository.