[#72] Remove code related to creating and updating repos (#143)
As a majority of the WATcher code base is ported over from CATcher,
it includes a lot of unused code. One such area is the code related to 
creating and updating issues on GitHub. Since WATcher is mainly used
to read data from GitHub and not modify data for now. 

Let's remove code that is related to modifying or creating repos