Improve activity dashboard design (#233)
Improve activity dashboard design

The ID column in the event table is not useful for the user.

The design of the dashboard is too simple. The event table's height is
not fixed, and the expanded event details might cause the table to be
too long, which is very inconvenient.

To address this, let's hide the ID column, add some information about
the actor, fix the event table height and make it scrollable, and
implement a dialog to display event details.