Enhance error message on incorrect organisation/repo name (#378)
The original error message for incorrect repo name does not specify 
whether the mistake is in the org name or repo name.

Hence, let's add an additional error message to show if the 
organisation name could not be found. This will give users an easier 
time finding the error and rectifying it when keying in a repo.