Handle templates using BehaviorSubject and async pipe (#407)
There is an error that appears when switching to an invalid repo. 
In the hidden-groups component, the templates are undefined before 
the view is initialized and they become defined afterwards, 
causing an NG0100 error. 

Let's use a BehaviorSubject with an initial value of null to emit the 
card template after the view has been initialized and use the async 
pipe so that the error does not occur.