Fix incorrect behaviour when filtering by assignee in issues with multiple assignees (#415)
Hide assignees not selected in filter

For issues with multiple assignees, having one of the assignees selected
in the filter displays all other assignees as groups.

This is misleading behaviour since the assignees that are not selected
are expected to be hidden.

Let's add a check for if the group is selected in the filter before
displaying the data.