Remove gap between bottom of window and horizontal sidebar (#420)
The height of each column in `card-view` component is set to 77vh,
which results in a gap between the bottom of the window and the
horizontal sidebar. 

The content should fill up the remaining height of the window instead
of leaving a gap.

Let's use a flexbox to allow the content to dynamically change its
height to fill up the remaining space instead of having a fixed height.