Update version deployment workflow to allow manual running (#432)
Update version deployment workflow

Version deployment is setup to run on push to deploy branch.

This is not consistent with CATcher's workflow, which is to manually
run the deployment from the actions tab on Github.

Let's update this to follow CATcher's workflow, changing it to be
manually run from the actions tab on Github.