Implement Presets for Filters (#411)
Add preset saving

Currently, there is no way to save a set of filters
for viewing. This may be inconvenient for e.g. tutors who need
to only manage a subset of people in the repository.

Let's enable users to create a "preset" of filters that they cann
activate whenever they want.