Remove back and home button
The back button and home button have ambiguous functions, and can
occasionally cause error messages to pop up.

This confuses the user, especially when they expect the back button to
bring them back to the repository selection page.

Let's remove the back and home buttons.

The back and home buttons are relics brought over from CATcher and do
not contribute significantly to WATcher's functionality. So, they can
and should be removed.