Show review decision and reviewers in PR card
Review decisions and reviewers are not shown in PR card.

Showing review decions and reviewers will allow users to see what needs
to be done and which PRs require reviews at a glance, and showing 
reviewers allows tutors to gauge the level of PR reviews in the team.
Displaying an icon for PRs that have been merged without reviews can 
help tutors to easily identify PRs that violate best practices for 
software engineering.

Let's ,
* display review decions in the PR card,
* display the GitHub avatars of PR reviewers in the PR card,
* use tooltips to show information about reviews, and
* display an icon for PRs that have been merged without any review.