Add a badge to indicate forking workflow compliance
There is currently no way to identify if a pull request is following
the forking workflow from the WATcher dashboard.

This makes it difficult for TAs to identify any pull requests that do
not follow the forking workflow at a glance.

Let's add a badge that indicates whether the pull request is following
the forking workflow.