Add link to GitHub profile from header card
Users have to manually type the URL with the assignee's GitHub
username to check their profiles.

This makes it tedious for users to look at an assignee's GitHub
profile to track issues and PRs contributed or other stuff.

Adding a direct link to the assignee's GitHub profile helps users
check an assignee's profile more conveniently.